check if status exists then delete

+++ app/controllers/benjamin_controller.rb
@@ -44,7 +44,7 @@ class BenjaminController < ApplicationController
   def user_profile(userid)
     url = "https://api.line.me/v2/bot/profile/#{userid}"
     profile = open(url, 'Authorization' => "Bearer #{ENV['LINE_TOKEN']}").read
-    profile.delete('statusMessage')
+    profile.delete('statusMessage') if profile['statusMessage']
     JSON.parse(user_profile)
   end
